Fix package structure: move controllers/dtos, move DBUserDetailsService, restore main class
