+++ src/main/java/edu/itc/enrollment_scheduling_system/controller/AdminController.java
@@ -1,5 +1,4 @@
-package edu.itc.enrollment_scheduling_system.web;
-
+package edu.itc.enrollment_scheduling_system.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
