+++ src/main/java/edu/itc/enrollment_scheduling_system/controller/AuthController.java
@@ -1,13 +1,12 @@
-package edu.itc.enrollment_scheduling_system.web;
 
+package edu.itc.enrollment_scheduling_system.controller;
 import edu.itc.enrollment_scheduling_system.service.RegistrationService;
-import edu.itc.enrollment_scheduling_system.web.dto.RegisterForm;
+import edu.itc.enrollment_scheduling_system.dto.RegisterForm;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-
 @Controller
 public class AuthController {
