+++ src/main/java/edu/itc/enrollment_scheduling_system/dto/RegisterForm.java
@@ -1,5 +1,4 @@
-package edu.itc.enrollment_scheduling_system.web.dto;
-
+package edu.itc.enrollment_scheduling_system.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
